@@ -1,6 +1,3 @@
-[bdist_wheel]
-universal = 1
-
 [bdist_rpm]
 requires = python python-urllib3
 
@@ -17,9 +14,7 @@ profile=black
 
 [black]
 max-line-length = 240
-target-version = 'py33'
+target-version = 'py38'
 
 [mypy]
 ignore_missing_imports=True
-
-
