@@ -26,13 +26,11 @@
 
 from typing import Any
 
-from six import string_types
-
 import opensearchpy
 from opensearchpy.serializer import serializer
 
 
-class Connections(object):
+class Connections:
     """
     Class responsible for holding connections to different clusters. Used as a
     singleton in this module.
@@ -106,7 +104,7 @@ class Connections(object):
         """
         # do not check isinstance(OpenSearch) so that people can wrap their
         # clients
-        if not isinstance(alias, string_types):
+        if not isinstance(alias, str):
             return alias
 
         # connection already established
