@@ -28,8 +28,6 @@ import collections.abc as collections_abc
 from fnmatch import fnmatch
 from typing import Any, Tuple, Type
 
-from six import add_metaclass, iteritems
-
 from opensearchpy.connection.connections import get_connection
 from opensearchpy.exceptions import NotFoundError, RequestError
 
@@ -41,7 +39,7 @@ from .search import Search
 from .utils import DOC_META_FIELDS, META_FIELDS, ObjectBase, merge
 
 
-class MetaField(object):
+class MetaField:
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         self.args, self.kwargs = args, kwargs
 
@@ -55,7 +53,7 @@ class DocumentMeta(type):
     ) -> Any:
         # DocumentMeta filters attrs in place
         attrs["_doc_type"] = DocumentOptions(name, bases, attrs)
-        return super(DocumentMeta, cls).__new__(cls, name, bases, attrs)
+        return super().__new__(cls, name, bases, attrs)
 
 
 class IndexMeta(DocumentMeta):
@@ -69,7 +67,7 @@ class IndexMeta(DocumentMeta):
         bases: Tuple[Type[ObjectBase]],
         attrs: Any,
     ) -> Any:
-        new_cls = super(IndexMeta, cls).__new__(cls, name, bases, attrs)
+        new_cls = super().__new__(cls, name, bases, attrs)
         if cls._document_initialized:
             index_opts = attrs.pop("Index", None)
             index = cls.construct_index(index_opts, bases)
@@ -96,7 +94,7 @@ class IndexMeta(DocumentMeta):
         return i
 
 
-class DocumentOptions(object):
+class DocumentOptions:
     def __init__(
         self,
         name: str,
@@ -109,7 +107,7 @@ class DocumentOptions(object):
         self.mapping = getattr(meta, "mapping", Mapping())
 
         # register all declared fields into the mapping
-        for name, value in list(iteritems(attrs)):
+        for name, value in list(attrs.items()):
             if isinstance(value, Field):
                 self.mapping.field(name, value)
                 del attrs[name]
@@ -130,8 +128,7 @@ class DocumentOptions(object):
         return self.mapping.properties.name
 
 
-@add_metaclass(DocumentMeta)
-class InnerDoc(ObjectBase):
+class InnerDoc(ObjectBase, metaclass=DocumentMeta):
     """
     Common class for inner documents like Object or Nested
     """
@@ -140,11 +137,10 @@ class InnerDoc(ObjectBase):
     def from_opensearch(cls, data: Any, data_only: bool = False) -> Any:
         if data_only:
             data = {"_source": data}
-        return super(InnerDoc, cls).from_opensearch(data)
+        return super().from_opensearch(data)
 
 
-@add_metaclass(IndexMeta)
-class Document(ObjectBase):
+class Document(ObjectBase, metaclass=IndexMeta):
     """
     Model-like class for persisting documents in opensearch.
     """
@@ -353,7 +349,7 @@ class Document(ObjectBase):
             ``[]``, ``{}``) to be left on the document. Those values will be
             stripped out otherwise as they make no difference in opensearch.
         """
-        d = super(Document, self).to_dict(skip_empty=skip_empty)
+        d = super().to_dict(skip_empty=skip_empty)
         if not include_meta:
             return d
 
