@@ -26,8 +26,6 @@
 from datetime import datetime, timedelta
 from typing import Any, Optional
 
-from six import iteritems, itervalues
-
 from opensearchpy.helpers.aggs import A
 
 from .query import MatchAll, Nested, Range, Terms
@@ -45,7 +43,7 @@ __all__ = [
 ]
 
 
-class Facet(object):
+class Facet:
     """
     A facet on faceted search. Wraps and aggregation and provides functionality
     to create a filter for selected values and return a list of facet values
@@ -146,7 +144,7 @@ class RangeFacet(Facet):
         return out
 
     def __init__(self, ranges: Any, **kwargs: Any) -> None:
-        super(RangeFacet, self).__init__(**kwargs)
+        super().__init__(**kwargs)
         self._params["ranges"] = list(map(self._range_to_dict, ranges))
         self._params["keyed"] = False
         self._ranges = dict(ranges)
@@ -217,7 +215,7 @@ class DateHistogramFacet(Facet):
 
     def __init__(self, **kwargs: Any) -> None:
         kwargs.setdefault("min_doc_count", 0)
-        super(DateHistogramFacet, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
     def get_value(self, bucket: Any) -> Any:
         if not isinstance(bucket["key"], datetime):
@@ -256,9 +254,7 @@ class NestedFacet(Facet):
     def __init__(self, path: Any, nested_facet: Any) -> None:
         self._path = path
         self._inner = nested_facet
-        super(NestedFacet, self).__init__(
-            path=path, aggs={"inner": nested_facet.get_aggregation()}
-        )
+        super().__init__(path=path, aggs={"inner": nested_facet.get_aggregation()})
 
     def get_values(self, data: Any, filter_values: Any) -> Any:
         return self._inner.get_values(data.inner, filter_values)
@@ -278,7 +274,7 @@ class FacetedResponse(Response):
     def facets(self) -> Any:
         if not hasattr(self, "_facets"):
             super(AttrDict, self).__setattr__("_facets", AttrDict({}))
-            for name, facet in iteritems(self._faceted_search.facets):
+            for name, facet in self._faceted_search.facets.items():
                 self._facets[name] = facet.get_values(
                     getattr(getattr(self.aggregations, "_filter_" + name), name),
                     self._faceted_search.filter_values.get(name, ()),
@@ -286,7 +282,7 @@ class FacetedResponse(Response):
         return self._facets
 
 
-class FacetedSearch(object):
+class FacetedSearch:
     """
     Abstraction for creating faceted navigation searches that takes care of
     composing the queries, aggregations and filters as needed as well as
@@ -344,7 +340,7 @@ class FacetedSearch(object):
         self._filters: Any = {}
         self._sort = sort
         self.filter_values: Any = {}
-        for name, value in iteritems(filters):
+        for name, value in filters.items():
             self.add_filter(name, value)
 
         self._s = self.build_search()
@@ -409,10 +405,10 @@ class FacetedSearch(object):
         Add aggregations representing the facets selected, including potential
         filters.
         """
-        for f, facet in iteritems(self.facets):
+        for f, facet in self.facets.items():
             agg = facet.get_aggregation()
             agg_filter = MatchAll()
-            for field, filter in iteritems(self._filters):
+            for field, filter in self._filters.items():
                 if f == field:
                     continue
                 agg_filter &= filter
@@ -429,7 +425,7 @@ class FacetedSearch(object):
             return search
 
         post_filter = MatchAll()
-        for f in itervalues(self._filters):
+        for f in self._filters.values():
             post_filter &= f
         return search.post_filter(post_filter)
 
