@@ -32,15 +32,13 @@ from datetime import date, datetime
 from typing import Any, Optional, Type
 
 from dateutil import parser, tz
-from six import integer_types, iteritems, string_types
-from six.moves import map
 
 from ..exceptions import ValidationException
 from .query import Q
 from .utils import AttrDict, AttrList, DslBase
 from .wrappers import Range
 
-unicode: Type[str] = type("")
+unicode: Type[str] = str
 
 
 def construct_field(name_or_field: Any, **params: Any) -> Any:
@@ -91,7 +89,7 @@ class Field(DslBase):
         """
         self._multi = multi
         self._required = required
-        super(Field, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def __getitem__(self, subfield: Any) -> Any:
         return self._params.get("fields", {})[subfield]
@@ -131,7 +129,7 @@ class Field(DslBase):
         return data
 
     def to_dict(self) -> Any:
-        d = super(Field, self).to_dict()
+        d = super().to_dict()
         name, value = d.popitem()
         value["type"] = name
         return value
@@ -145,7 +143,7 @@ class CustomField(Field):
         if isinstance(self.builtin_type, Field):
             return self.builtin_type.to_dict()
 
-        d = super(CustomField, self).to_dict()
+        d = super().to_dict()
         d["type"] = self.builtin_type
         return d
 
@@ -183,13 +181,13 @@ class Object(Field):
 
             # no InnerDoc subclass, creating one instead...
             self._doc_class = type("InnerDoc", (InnerDoc,), {})
-            for name, field in iteritems(properties or {}):
+            for name, field in (properties or {}).items():
                 self._doc_class._doc_type.mapping.field(name, field)  # type: ignore
             if dynamic is not None:
                 self._doc_class._doc_type.mapping.meta("dynamic", dynamic)  # type: ignore
 
         self._mapping = copy.deepcopy(self._doc_class._doc_type.mapping)
-        super(Object, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
     def __getitem__(self, name: Any) -> Any:
         return self._mapping[name]
@@ -210,7 +208,7 @@ class Object(Field):
 
     def to_dict(self) -> Any:
         d = self._mapping.to_dict()
-        d.update(super(Object, self).to_dict())
+        d.update(super().to_dict())
         return d
 
     def _collect_fields(self) -> Any:
@@ -237,7 +235,7 @@ class Object(Field):
         return data.to_dict()
 
     def clean(self, data: Any) -> Any:
-        data = super(Object, self).clean(data)
+        data = super().clean(data)
         if data is None:
             return None
         if isinstance(data, (list, AttrList)):
@@ -260,7 +258,7 @@ class Nested(Object):
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         kwargs.setdefault("multi", True)
-        super(Nested, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
 
 class Date(Field):
@@ -273,12 +271,12 @@ class Date(Field):
             May be instance of `datetime.tzinfo` or string containing TZ offset
         """
         self._default_timezone = default_timezone
-        if isinstance(self._default_timezone, string_types):
+        if isinstance(self._default_timezone, str):
             self._default_timezone = tz.gettz(self._default_timezone)
-        super(Date, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
 
     def _deserialize(self, data: Any) -> Any:
-        if isinstance(data, string_types):
+        if isinstance(data, str):
             try:
                 data = parser.parse(data)
             except Exception as e:
@@ -292,7 +290,7 @@ class Date(Field):
             return data
         if isinstance(data, date):
             return data
-        if isinstance(data, integer_types):
+        if isinstance(data, int):
             # Divide by a float to preserve milliseconds on the datetime.
             return datetime.utcfromtimestamp(data / 1000.0)
 
@@ -361,7 +359,7 @@ class DenseVector(Float):
 
     def __init__(self, dims: Any, **kwargs: Any) -> None:
         kwargs["multi"] = True
-        super(DenseVector, self).__init__(dims=dims, **kwargs)
+        super().__init__(dims=dims, **kwargs)
 
 
 class SparseVector(Field):
@@ -376,9 +374,7 @@ class ScaledFloat(Float):
     name: Optional[str] = "scaled_float"
 
     def __init__(self, scaling_factor: Any, *args: Any, **kwargs: Any) -> None:
-        super(ScaledFloat, self).__init__(
-            scaling_factor=scaling_factor, *args, **kwargs
-        )
+        super().__init__(scaling_factor=scaling_factor, *args, **kwargs)
 
 
 class Double(Float):
@@ -481,7 +477,7 @@ class RangeField(Field):
     def _deserialize(self, data: Any) -> Any:
         if isinstance(data, Range):
             return data
-        data = dict((k, self._core_field.deserialize(v)) for k, v in iteritems(data))
+        data = {k: self._core_field.deserialize(v) for k, v in data.items()}
         return Range(data)
 
     def _serialize(self, data: Any) -> Any:
@@ -489,7 +485,7 @@ class RangeField(Field):
             return None
         if not isinstance(data, collections_abc.Mapping):
             data = data.to_dict()
-        return dict((k, self._core_field.serialize(v)) for k, v in iteritems(data))
+        return {k: self._core_field.serialize(v) for k, v in data.items()}
 
 
 class IntegerRange(RangeField):
