@@ -24,14 +24,10 @@
 #  specific language governing permissions and limitations
 #  under the License.
 
-from __future__ import unicode_literals
 
 import collections.abc as collections_abc
 from copy import copy
 from typing import Any, Callable, Dict, Optional, Tuple
-
-from six import add_metaclass, iteritems
-from six.moves import map
 
 from opensearchpy.exceptions import UnknownDslObject, ValidationException
 
@@ -66,7 +62,7 @@ def _wrap(val: Any, obj_wrapper: Optional[Callable[..., Any]] = None) -> Any:
     return val
 
 
-class AttrList(object):
+class AttrList:
     def __init__(
         self, p: Any, obj_wrapper: Optional[Callable[..., Any]] = None
     ) -> None:
@@ -118,7 +114,7 @@ class AttrList(object):
         self._l_, self._obj_wrapper = state
 
 
-class AttrDict(object):
+class AttrDict:
     """
     Helper class to provide attribute like access (read and write) to
     dictionaries. Used to provide a convenient way to access both results and
@@ -127,7 +123,7 @@ class AttrDict(object):
 
     def __init__(self, d: Any) -> None:
         # assign the inner dict manually to prevent __setattr__ from firing
-        super(AttrDict, self).__setattr__("_d_", d)
+        super().__setattr__("_d_", d)
 
     def __contains__(self, key: Any) -> bool:
         return key in self._d_
@@ -160,7 +156,7 @@ class AttrDict(object):
         return (self._d_,)
 
     def __setstate__(self, state: Any) -> None:
-        super(AttrDict, self).__setattr__("_d_", state[0])
+        super().__setattr__("_d_", state[0])
 
     def __getattr__(self, attr_name: Any) -> Any:
         try:
@@ -204,7 +200,7 @@ class AttrDict(object):
             self._d_[name] = value
         else:
             # there is an attribute on the class (could be property, ..) - don't add it as field
-            super(AttrDict, self).__setattr__(name, value)
+            super().__setattr__(name, value)
 
     def __iter__(self) -> Any:
         return iter(self._d_)
@@ -230,7 +226,7 @@ class DslMeta(type):
 
     def __init__(cls: Any, name: str, bases: Any, attrs: Any) -> None:
         # TODO: why is it calling itself?!
-        super(DslMeta, cls).__init__(name, bases, attrs)
+        super().__init__(name, bases, attrs)
         # skip for DslBase
         if not hasattr(cls, "_type_shortcut"):
             return
@@ -252,8 +248,7 @@ class DslMeta(type):
             raise UnknownDslObject("DSL type %s does not exist." % name)
 
 
-@add_metaclass(DslMeta)
-class DslBase(object):
+class DslBase(metaclass=DslMeta):
     """
     Base class for all DSL objects - queries, filters, aggregations etc. Wraps
     a dictionary representing the object's json.
@@ -285,7 +280,7 @@ class DslBase(object):
 
     def __init__(self, _expand__to_dot: Any = EXPAND__TO_DOT, **params: Any) -> None:
         self._params = {}
-        for pname, pvalue in iteritems(params):
+        for pname, pvalue in params.items():
             if "__" in pname and _expand__to_dot:
                 pname = pname.replace("__", ".")
             self._setattr(pname, pvalue)
@@ -294,7 +289,7 @@ class DslBase(object):
         """Produce a repr of all our parameters to be used in __repr__."""
         return ", ".join(
             "{}={!r}".format(n.replace(".", "__"), v)
-            for (n, v) in sorted(iteritems(self._params))
+            for (n, v) in sorted(self._params.items())
             # make sure we don't include empty typed params
             if "type" not in self._param_defs.get(n, {}) or v
         )
@@ -310,7 +305,7 @@ class DslBase(object):
 
     def __setattr__(self, name: str, value: Any) -> None:
         if name.startswith("_"):
-            return super(DslBase, self).__setattr__(name, value)
+            return super().__setattr__(name, value)
         return self._setattr(name, value)
 
     def _setattr(self, name: Any, value: Any) -> None:
@@ -327,7 +322,7 @@ class DslBase(object):
                     if not isinstance(value, (tuple, list)):
                         value = (value,)
                     value = list(
-                        {k: shortcut(v) for (k, v) in iteritems(obj)} for obj in value
+                        {k: shortcut(v) for (k, v) in obj.items()} for obj in value
                     )
                 elif pinfo.get("multi"):
                     if not isinstance(value, (tuple, list)):
@@ -336,7 +331,7 @@ class DslBase(object):
 
                 # dict(name -> DslBase), make sure we pickup all the objs
                 elif pinfo.get("hash"):
-                    value = {k: shortcut(v) for (k, v) in iteritems(value)}
+                    value = {k: shortcut(v) for (k, v) in value.items()}
 
                 # single value object, just convert
                 else:
@@ -380,7 +375,7 @@ class DslBase(object):
         Serialize the DSL object to plain dict
         """
         d = {}
-        for pname, value in iteritems(self._params):
+        for pname, value in self._params.items():
             pinfo = self._param_defs.get(pname)
 
             # typed param
@@ -392,7 +387,7 @@ class DslBase(object):
                 # list of dict(name -> DslBase)
                 if pinfo.get("multi") and pinfo.get("hash"):
                     value = list(
-                        {k: v.to_dict() for k, v in iteritems(obj)} for obj in value
+                        {k: v.to_dict() for k, v in obj.items()} for obj in value
                     )
 
                 # multi-values are serialized as list of dicts
@@ -401,7 +396,7 @@ class DslBase(object):
 
                 # squash all the hash values into one dict
                 elif pinfo.get("hash"):
-                    value = {k: v.to_dict() for k, v in iteritems(value)}
+                    value = {k: v.to_dict() for k, v in value.items()}
 
                 # serialize single values
                 else:
@@ -427,13 +422,13 @@ class HitMeta(AttrDict):
     ) -> None:
         d = {
             k[1:] if k.startswith("_") else k: v
-            for (k, v) in iteritems(document)
+            for (k, v) in document.items()
             if k not in exclude
         }
         if "type" in d:
             # make sure we are consistent everywhere in python
             d["doc_type"] = d.pop("type")
-        super(HitMeta, self).__init__(d)
+        super().__init__(d)
 
 
 class ObjectBase(AttrDict):
@@ -447,7 +442,7 @@ class ObjectBase(AttrDict):
 
         super(AttrDict, self).__setattr__("meta", HitMeta(meta))
 
-        super(ObjectBase, self).__init__(kwargs)
+        super().__init__(kwargs)
 
     @classmethod
     def __list_fields(cls: Any) -> Any:
@@ -491,7 +486,7 @@ class ObjectBase(AttrDict):
         return doc
 
     def _from_dict(self, data: Any) -> None:
-        for k, v in iteritems(data):
+        for k, v in data.items():
             f = self.__get_field(k)
             if f and f._coerce:
                 v = f.deserialize(v)
@@ -508,7 +503,7 @@ class ObjectBase(AttrDict):
 
     def __getattr__(self, name: Any) -> Any:
         try:
-            return super(ObjectBase, self).__getattr__(name)
+            return super().__getattr__(name)
         except AttributeError:
             f = self.__get_field(name)
             if hasattr(f, "empty"):
@@ -521,7 +516,7 @@ class ObjectBase(AttrDict):
 
     def to_dict(self, skip_empty: Optional[bool] = True) -> Any:
         out = {}
-        for k, v in iteritems(self._d_):
+        for k, v in self._d_.items():
             # if this is a mapped field,
             f = self.__get_field(k)
             if f and f._coerce:
@@ -584,7 +579,7 @@ def merge(data: Any, new_data: Any, raise_on_conflict: bool = False) -> None:
             )
         )
 
-    for key, value in iteritems(new_data):
+    for key, value in new_data.items():
         if (
             key in data
             and isinstance(data[key], (AttrDict, collections_abc.Mapping))
